+++ scanpipe/tests/test_pipes.py
@@ -1135,10 +1135,12 @@ class ScanPipePipesTransactionTest(TransactionTestCase):
         p1 = Project.objects.create(name="Analysis")
         resource_location = str(self.data_location / "notice.NOTICE")
 
-        with self.assertRaises(AssertionError) as cm:
+        with self.assertRaises(ValueError) as cm:
             make_codebase_resource(p1, resource_location)
 
-        self.assertIn("is not under project/codebase/", str(cm.exception))
+        self.assertIn("not", str(cm.exception))
+        self.assertIn(resource_location, str(cm.exception))
+        self.assertIn("/codebase", str(cm.exception))
 
         copy_input(resource_location, p1.codebase_path)
         resource_location = str(p1.codebase_path / "notice.NOTICE")